// gemip-mipcittadinowcl/src/app/integration/model/domanda.ts
/**
 * @license
 *
 * Copyright © 2025 Regione Piemonte
 *
 * Licensed under the EUPL, Version 1.2 or – as soon they will be
 * approved by the European Commission - subsequent versions of the
 * EUPL (the "Licence");
 *
 * You may not use this work except in compliance with the Licence.
 * You may obtain a copy of the Licence at:
 *
 * https://interoperable-europe.ec.europa.eu/collection/eupl/eupl-text-eupl-12
 * or
 * https://opensource.org/license/EUPL-1.2
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the Licence is distributed on an "AS IS" basis,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the Licence for the specific language governing permissions and
 * limitations under the Licence.
 */

/**
 * API mipcommonapi
 * API comuni mipcittadino e mipbackoffice
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { Risposta } from './risposta';

export interface Domanda { 
    idDomanda: number;
    testoDomanda: string;
    placeholder?: string;
    tipoDomanda: Domanda.TipoDomandaEnum;
    risposte: Array<Risposta>;
    idDomandaCondizionale?: number;
    idRispostaCondizionale?: number;
}
export namespace Domanda {
    export type TipoDomandaEnum = 'aperta' | 'chiusa' | 'chiusa_scelta_multipla';
    export const TipoDomandaEnum = {
        Aperta: 'aperta' as TipoDomandaEnum,
        Chiusa: 'chiusa' as TipoDomandaEnum,
        ChiusaSceltaMultipla: 'chiusa_scelta_multipla' as TipoDomandaEnum
    };
}